libgrit: Expand size of attribute fields of GRF file
For example, this is required so that the palette size can be 256, which
doesn't fit in 8 bits.

+++ libgrit/grit_xp.cpp
@@ -126,9 +126,9 @@ struct chunk_t
 struct GrfHeader
 {
 	union {
-		u8 attrs[4];
+		u16 attrs[4];
 		struct {
-			u8	gfxAttr, mapAttr, mmapAttr, palAttr;
+			u16	gfxAttr, mapAttr, mmapAttr, palAttr;
 		};
 	};
 	u8		tileWidth, tileHeight;
